+++ ignite/services/chain/serve.go
@@ -17,8 +17,12 @@ import (
 	"github.com/ignite/cli/ignite/chainconfig"
 	"github.com/ignite/cli/ignite/pkg/cache"
 	chaincmdrunner "github.com/ignite/cli/ignite/pkg/chaincmd/runner"
+	"github.com/ignite/cli/ignite/pkg/cliui/colors"
+	"github.com/ignite/cli/ignite/pkg/cliui/icons"
+	"github.com/ignite/cli/ignite/pkg/cliui/view/errorview"
 	"github.com/ignite/cli/ignite/pkg/cosmosfaucet"
 	"github.com/ignite/cli/ignite/pkg/dirchange"
+	"github.com/ignite/cli/ignite/pkg/events"
 	"github.com/ignite/cli/ignite/pkg/localfs"
 	"github.com/ignite/cli/ignite/pkg/xexec"
 	"github.com/ignite/cli/ignite/pkg/xfilepath"
@@ -125,6 +129,7 @@ func (c *Chain) Serve(ctx context.Context, cacheStorage cache.Storage, options .
 			if ctx.Err() != nil {
 				return ctx.Err()
 			}
+
 			select {
 			case <-ctx.Done():
 				return ctx.Err()
@@ -136,10 +141,12 @@ func (c *Chain) Serve(ctx context.Context, cacheStorage cache.Storage, options .
 				}
 
 				var (
-					serveCtx context.Context
-					buildErr *CannotBuildAppError
-					startErr *CannotStartAppError
+					serveCtx      context.Context
+					buildErr      *CannotBuildAppError
+					startErr      *CannotStartAppError
+					validationErr *chainconfig.ValidationError
 				)
+
 				serveCtx, c.serveCancel = context.WithCancel(ctx)
 
 				// determine if the chain should reset the state
@@ -156,31 +163,28 @@ func (c *Chain) Serve(ctx context.Context, cacheStorage cache.Storage, options .
 					if c.served {
 						c.served = false
 
-						fmt.Fprintln(c.stdLog().out, "💿 Saving genesis state...")
+						c.ev.Send("💿 Saving genesis state...")
 
 						// If serve has been stopped, save the genesis state
 						if err := c.saveChainState(context.TODO(), commands); err != nil {
-							fmt.Fprint(c.stdLog().err, err.Error())
+							c.ev.SendError(err)
 							return err
 						}
 
 						genesisPath, err := c.exportedGenesisPath()
 						if err != nil {
-							fmt.Fprintln(c.stdLog().err, err.Error())
 							return err
 						}
-						fmt.Fprintf(c.stdLog().out, "💿 Genesis state saved in %s\n", genesisPath)
+
+						c.ev.Sendf("💿 Genesis state saved in %s", genesisPath)
 					}
+				case errors.As(err, &validationErr):
+					// Change error message to add a link to the configuration docs
+					err = fmt.Errorf("%w\nsee: https://github.com/ignite/cli#configure", err)
+
+					c.ev.SendView(errorview.NewError(err), events.ProgressFinished())
 				case errors.As(err, &buildErr):
-					fmt.Fprintf(c.stdLog().err, "%s\n", errorColor(err.Error()))
-
-					var validationErr *chainconfig.ValidationError
-					if errors.As(err, &validationErr) {
-						fmt.Fprintln(c.stdLog().out, "see: https://github.com/ignite/cli#configure")
-					}
-
-					fmt.Fprintf(c.stdLog().out, "%s\n", infoColor("Waiting for a fix before retrying..."))
-
+					c.ev.SendView(errorview.NewError(err), events.ProgressFinished())
 				case errors.As(err, &startErr):
 					// Parse returned error logs
 					parsedErr := startErr.ParseStartError()
@@ -189,9 +193,13 @@ func (c *Chain) Serve(ctx context.Context, cacheStorage cache.Storage, options .
 					// Therefore, the error may be caused by a new logic that is not compatible with the old app state
 					// We suggest the user to eventually reset the app state
 					if parsedErr == "" {
-						fmt.Fprintf(c.stdLog().out, "%s %s\n", infoColor(`Blockchain failed to start.
-If the new code is no longer compatible with the saved state, you can reset the database by launching:`), "ignite chain serve --reset-once")
-
+						c.ev.Send(
+							fmt.Sprintf(
+								"%s %s",
+								colors.Info("Blockchain failed to start.\nIf the new code is no longer compatible with the saved state, you can reset the database by launching:"),
+								"ignite chain serve --reset-once",
+							),
+						)
 						return fmt.Errorf("cannot run %s", startErr.AppName)
 					}
 
@@ -213,7 +221,7 @@ If the new code is no longer compatible with the saved state, you can reset the 
 }
 
 func (c *Chain) setup() error {
-	fmt.Fprintf(c.stdLog().out, "Cosmos SDK's version is: %s\n\n", infoColor(c.Version))
+	c.ev.Sendf("Cosmos SDK's version is: %s\n", colors.Info(c.Version))
 
 	return c.checkSystem()
 }
@@ -232,6 +240,7 @@ func (c *Chain) refreshServe() {
 	if c.serveCancel != nil {
 		c.serveCancel()
 	}
+	// send event changes detected
 	c.serveRefresher <- struct{}{}
 }
 
@@ -288,7 +297,7 @@ func (c *Chain) serve(ctx context.Context, cacheStorage cache.Storage, forceRese
 
 		if forceReset || configModified {
 			// if forceReset is set, we consider the app as being not initialized
-			fmt.Fprintln(c.stdLog().out, "🔄 Resetting the app state...")
+			c.ev.Send("🔄 Resetting the app state...")
 			isInit = false
 		}
 	}
@@ -344,7 +353,7 @@ func (c *Chain) serve(ctx context.Context, cacheStorage cache.Storage, forceRese
 	// init phase
 	// nolint:gocritic
 	if !isInit || (appModified && !exportGenesisExists) {
-		fmt.Fprintln(c.stdLog().out, "💿 Initializing the app...")
+		c.ev.Send("💿 Initializing the app...")
 
 		if err := c.Init(ctx, true); err != nil {
 			return err
@@ -352,7 +361,7 @@ func (c *Chain) serve(ctx context.Context, cacheStorage cache.Storage, forceRese
 	} else if appModified {
 		// if the chain is already initialized but the source has been modified
 		// we reset the chain database and import the genesis state
-		fmt.Fprintln(c.stdLog().out, "💿 Existent genesis detected, restoring the database...")
+		c.ev.Send("💿 Existent genesis detected, restoring the database...")
 
 		if err := commands.UnsafeReset(ctx); err != nil {
 			return err
@@ -362,7 +371,7 @@ func (c *Chain) serve(ctx context.Context, cacheStorage cache.Storage, forceRese
 			return err
 		}
 	} else {
-		fmt.Fprintln(c.stdLog().out, "▶️  Restarting existing app...")
+		c.ev.Send("▶  Restarting existing app...")
 	}
 
 	// save checksums
@@ -433,13 +442,23 @@ func (c *Chain) start(ctx context.Context, config *chainconfig.Config) error {
 	rpcAddr, _ := xurl.HTTP(servers.RPC.Address)
 	apiAddr, _ := xurl.HTTP(servers.API.Address)
 
-	// print the server addresses.
-	fmt.Fprintf(c.stdLog().out, "🌍 Tendermint node: %s\n", rpcAddr)
-	fmt.Fprintf(c.stdLog().out, "🌍 Blockchain API: %s\n", apiAddr)
+	c.ev.Send(
+		fmt.Sprintf("Tendermint node: %s", rpcAddr),
+		events.Icon(icons.Earth),
+		events.ProgressFinished(),
+	)
+	c.ev.Send(
+		fmt.Sprintf("Blockchain API: %s", apiAddr),
+		events.Icon(icons.Earth),
+	)
 
 	if isFaucetEnabled {
 		faucetAddr, _ := xurl.HTTP(chainconfig.FaucetHost(config))
-		fmt.Fprintf(c.stdLog().out, "🌍 Token faucet: %s\n", faucetAddr)
+
+		c.ev.Send(
+			fmt.Sprintf("Token faucet: %s", faucetAddr),
+			events.Icon(icons.Earth),
+		)
 	}
 
 	return g.Wait()