+++ ignite/services/network/campaign.go
@@ -47,7 +47,7 @@ func WithCampaignTotalSupply(totalSupply sdk.Coins) Prop {
 
 // Campaign fetches the campaign from Network
 func (n Network) Campaign(ctx context.Context, campaignID uint64) (networktypes.Campaign, error) {
-	n.ev.Send(events.New(events.StatusOngoing, "Fetching campaign information"))
+	n.ev.Send("Fetching campaign information", events.ProgressStarted())
 	res, err := n.campaignQuery.Campaign(ctx, &campaigntypes.QueryGetCampaignRequest{
 		CampaignID: campaignID,
 	})
@@ -63,9 +63,8 @@ func (n Network) Campaign(ctx context.Context, campaignID uint64) (networktypes.
 func (n Network) Campaigns(ctx context.Context) ([]networktypes.Campaign, error) {
 	var campaigns []networktypes.Campaign
 
-	n.ev.Send(events.New(events.StatusOngoing, "Fetching campaigns information"))
-	res, err := n.campaignQuery.
-		CampaignAll(ctx, &campaigntypes.QueryAllCampaignRequest{})
+	n.ev.Send("Fetching campaigns information", events.ProgressStarted())
+	res, err := n.campaignQuery.CampaignAll(ctx, &campaigntypes.QueryAllCampaignRequest{})
 	if err != nil {
 		return campaigns, err
 	}
@@ -80,11 +79,12 @@ func (n Network) Campaigns(ctx context.Context) ([]networktypes.Campaign, error)
 
 // CreateCampaign creates a campaign in Network
 func (n Network) CreateCampaign(ctx context.Context, name, metadata string, totalSupply sdk.Coins) (uint64, error) {
-	n.ev.Send(events.New(events.StatusOngoing, fmt.Sprintf("Creating campaign %s", name)))
+	n.ev.Send(fmt.Sprintf("Creating campaign %s", name), events.ProgressStarted())
 	addr, err := n.account.Address(networktypes.SPN)
 	if err != nil {
 		return 0, err
 	}
+
 	msgCreateCampaign := campaigntypes.NewMsgCreateCampaign(
 		addr,
 		name,
@@ -112,7 +112,7 @@ func (n Network) InitializeMainnet(
 	sourceHash string,
 	mainnetChainID string,
 ) (uint64, error) {
-	n.ev.Send(events.New(events.StatusOngoing, "Initializing the mainnet campaign"))
+	n.ev.Send("Initializing the mainnet campaign", events.ProgressStarted())
 	addr, err := n.account.Address(networktypes.SPN)
 	if err != nil {
 		return 0, err
@@ -136,7 +136,7 @@ func (n Network) InitializeMainnet(
 		return 0, err
 	}
 
-	n.ev.Send(events.New(events.StatusDone, fmt.Sprintf("Campaign %d initialized on mainnet", campaignID)))
+	n.ev.Send(fmt.Sprintf("Campaign %d initialized on mainnet", campaignID), events.ProgressFinished())
 
 	return initMainnetRes.MainnetID, nil
 }
@@ -153,11 +153,12 @@ func (n Network) UpdateCampaign(
 		apply(&p)
 	}
 
-	n.ev.Send(events.New(events.StatusOngoing, fmt.Sprintf("Updating the campaign %d", id)))
+	n.ev.Send(fmt.Sprintf("Updating the campaign %d", id), events.ProgressStarted())
 	account, err := n.account.Address(networktypes.SPN)
 	if err != nil {
 		return err
 	}
+
 	msgs := make([]sdk.Msg, 0)
 	if p.name != "" || len(p.metadata) > 0 {
 		msgs = append(msgs, campaigntypes.NewMsgEditCampaign(
@@ -178,8 +179,6 @@ func (n Network) UpdateCampaign(
 	if _, err := n.cosmos.BroadcastTx(ctx, n.account, msgs...); err != nil {
 		return err
 	}
-	n.ev.Send(events.New(events.StatusDone, fmt.Sprintf(
-		"Campaign %d updated", id,
-	)))
+	n.ev.Send(fmt.Sprintf("Campaign %d updated", id), events.ProgressFinished())
 	return nil
 }