+++ changelog.md
@@ -12,6 +12,7 @@
 - Updated `nodetime`: `ts-proto` to `v1.123.0`, `protobufjs` to `v7.1.1`, `swagger-typescript-api` to `v9.2.0`
 - Switched codegen client to use `axios` instead of `fetch`
 - nodetime built with `vercel/pkg@5.6.0`
+- Change CLI to use an events bus to print to stdout.
 - Move generated proto files to `proto/{appname}/{module}`
 
 ### Fixes